feat: mypy hardening sweep — eliminate ~1,800 Any usages and standardise tempo to int
## What changed

### New type contracts
- `app/contracts/generation_types.py`: GenerationContext, CompositionContext (+ role/previous_notes/drum_telemetry fields), RoleResult, UnifiedGenerationOutput
- `app/contracts/llm_types.py`: OpenAIData quarantine alias + OpenAIStreamChunk, OpenAITool, UsageStats, OpenAIRequestPayload, OpenAIResponse, OpenAIToolChoice, StreamEvent, ChatMessage, ToolCallEntry, SystemMessage/UserMessage/AssistantMessage/ToolResultMessage
- `app/contracts/json_types.py`, `app/contracts/project_types.py`, `app/contracts/mcp_types.py`: migrated from ad-hoc dicts
- `app/auth/tokens.py`: TokenClaims TypedDict (type/iat/exp/sub/role) + propagate role claim through validate_access_code
- `app/services/orpheus.py`: OrpheusRawResponse TypedDict
- `app/services/muse_drift.py`: SampleChange TypedDict
- `app/core/planner/conversion.py`: _ExistingTrackInfo TypedDict
- `app/core/state_store.py`: _ProjectMetadataSnapshot TypedDict

### Orpheus types (orpheus-music/orpheus_types.py)
- OrpheusNoteDict, OrpheusCCEvent, OrpheusPitchBend, OrpheusAftertouch (MIDI events)
- ParsedMidiResult, WireNoteDict, CacheKeyData, FulfillmentReport, GradioGenerationParams
- GenerationComparison, QualityEvalParams, QualityEvalToolCall
- ScoringParams @dataclass, BestCandidate @dataclass

### Tempo standardised to int
- Beats per minute is a whole integer by definition — float was an API misdesign
- Single coercion point: `int(round(float(...)))` in coordinator.py at the DAW boundary
- ProjectContext, CompositionContext, all contracts, SectionTelemetry, CacheKeyData now use int

### Any quarantine
- `Any` lives in exactly two places: llm_types.py (OpenAIData) and orpheus boundary
- All other modules import named aliases; no naked dict[str, Any] in production code
- GenerationResult.metadata widened to dict[str, object] for genuine flexibility

### Documentation
- `docs/reference/type-contracts.md`: full API surface of every entity with field tables, descriptions, the tempo convention, the Any quarantine strategy, and the full entity hierarchy

### Tooling
- `tools/typing_audit.py`: AST-based script to count Any patterns by file for tracking progress

mypy: Success 368 source files | tests: 3704 passed + 221 passed (orpheus)

+++ app/protocol/schemas/project.py
@@ -81,7 +81,7 @@ class ProjectSnapshot(CamelModel):
 
     id: str
     name: str | None = None
-    tempo: float | None = Field(default=None, ge=40, le=240)
+    tempo: int | None = Field(default=None, ge=40, le=240)
     key: str | None = None
     time_signature: str | None = None
     schema_version: int | None = None
